Move BuildDebugSceneGraph function, update all calls to this function
